ref: initial engine to ref API draft, not compiling, work in progress [ci skip]
