+++ ref_gl/gl_rpart.c
@@ -79,7 +79,7 @@ void CL_DrawParticles( double frametime, particle_t *cl_active_particles )
 			p->color = bound( 0, p->color, 255 );
 			pColor = &clgame.palette[p->color];
 
-			alpha = 255 * (p->die - cl.time) * 16.0f;
+			alpha = 255 * (p->die - gpGlobals->time) * 16.0f;
 			if( alpha > 255 || p->type == pt_static )
 				alpha = 255;
 
@@ -258,7 +258,7 @@ void CL_DrawTracers( double frametime, particle_t *cl_active_tracers )
 
 	for( p = cl_active_tracers; p; p = p->next )
 	{
-		atten = (p->die - cl.time);
+		atten = (p->die - gpGlobals->time);
 		if( atten > 0.1f ) atten = 0.1f;
 
 		VectorScale( p->vel, ( p->ramp * atten ), delta );
@@ -317,7 +317,7 @@ void CL_DrawTracers( double frametime, particle_t *cl_active_tracers )
 			p->vel[1] *= scale;
 			p->vel[2] -= gravity;
 
-			p->packedColor = 255 * (p->die - cl.time) * 2;
+			p->packedColor = 255 * (p->die - gpGlobals->time) * 2;
 			if( p->packedColor > 255 ) p->packedColor = 255;
 		}
 		else if( p->type == pt_slowgrav )
