+++ ref_gl/wscript
@@ -20,6 +20,8 @@ def configure(conf):
 	if conf.options.SUPPORT_BSP2_FORMAT:
 		conf.env.append_unique('DEFINES', 'SUPPORT_BSP2_FORMAT')
 
+	conf.env.append_unique('DEFINES', 'REF_DLL')
+
 def build(bld):
 	if bld.env.DEDICATED:
 		return
@@ -41,7 +43,7 @@ def build(bld):
 		'../pm_shared' ]
 
 	bld.shlib(
-	    source   = source,
+		source   = source,
 		target   = name,
 		features = 'c',
 		includes = includes,
